+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # Change log
 
+## 2.7.5 (release date: 23.11.2020)
+
+### Framework
+* Fix playback with start position causing transition errors on PC
+
+### Tools
+* Fix compatibility with PostCSS 8.0
+* Closure compiler updated to `20201102`
+
+## 2.7.4 (release date: 01.10.2020)
+Virtually the same as `2.7.4-alpha.1`
+
+### Framework
+* Fix start position not being applied on webOS in some cases
+
+### Tools
+* Closure compiler updated to `20200927`
+
+## 2.7.4-alpha.1 (release date: 30.09.2020)
+
+### Framework
+* Experimental fix for an issue on webOS 2017 when start position fails to apply in some cases.
+
+## 2.7.3 (release date: 05.08.2020)
+
+### Framework
+* Added export if `NativeReadyState` to `StatefulHtml5Video`
+
+### Tools
+* Replaced `chalk` with `kleur`
+* Improved output of `--help` option for CLI tools
+
 ## 2.7.2 (release date: 22.04.2020)
 
 ### Framework